Create 1.c
In the file , two functions, int_to_linked_list and free_linked_list, are already implemented.

The function int_to_linked_list stores the individual digits of a given positive integer greater than or equal to 0 in a linked list. Starting from the head node of the list, the values of the nodes represent the digits of the number read from right to left.

The function free_linked_list deallocates the memory reserved for the linked list.

The task involves understanding how to work with lists, recursion, and iteration, as well as ensuring correct processing and output of data.